Fix UTF-8 fuckup in titleshortening.
Since the offset may fall exactly between a multibytecharacter, which then gets
ripped apart, we need to use a special function(toUTFindex) to calculate the offset,
where in the characterarray the title_cutoff-th character is.

This introduces a certain vagueness, how long the shortened string will be.
If the cont string has a lot of multibytecharacters, the title_cutoff will be matched sooner,
so the title will be shorter.