Add captcha solving support.
Facebook thought it a nifty "security" feature, to add
"security check"-captcha to their pages.

Whilst this has nothing to do with "security", it effectively renders
my attempts at fetching the data useless.

So here's what to do:
1) Add cookie storage to the request
2) Solve a captcha once
3) ???
4) Non-Profit

Which is what was implemented in this commit.
You'll have to execute the ./captcha-Binary once, and interpret the string
inside the image, and then it should continue to work normally.

For compatibility-reasons I included the standardpaths-library, which handles
Paths quite nicely.